requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
python-multipart
mangum